models/requirements.txt
notebook==6.4.4
numpy==1.21.2
scikit-learn==0.24.2
pandas==1.3.3 
data-science-types==0.2.23
python-dotenv==0.19.0
torch==1.9.0
torchmetrics==0.5.1
pytorch-lightning==1.4.7
hydra-core==1.1.1
transformers==4.10.3
hdbscan==0.8.27 # Unupervised clustering 
requests-futures==1.0.0 # Async http requests 